Confused bug fix
Found that addSample in csv_parser wasn't returning a value so would always fail
Not certain this was the source of the problem, but this definitely fixed something

Also discovered DB Browser for SQLite, so I can view table contents in a visual
table on my desktop. Loads easier. CSV import now definitely only inserting 6941
results, which I've verified in Excel.
https://sqlitebrowser.org/dl/