Back to Drop Off Point
Okay, I didn't do a lot of useful commits. This change started as a simplification of the
database schema to enable rapid turnaround of queries compared to using three linked
tables etc. That turned into a problem solving issue for the csv_parser, which turned
into a speed optimisation of the same script. That lead to the addition of indexes on
the database and some rewrites of the sql strings to massively improve execution time
(3000ms down to 1/2ms in a lot of cases). Then once the linear regression script needed
updating in a similar way, until I finally got to this point, where I can get a plot of
the linear regression displayed with colour banding and all that jazz. Something of note:
sqlite didn't like me passing 0.0011... to an escaped insert ? to a REAL field. Not sure
why, but it converted ALL results sent to it, by almost but not quite double, 0.0022...
This meant that the plots displayed telepath points in one region, and the linear regress
in another. Took ages to troubleshoot that one.

Anyway, sorry for bad commit habbits. Hopefully back to normal now.
finally, I've been able to