CSV Parser Complete
Reordered processResults to do the dob/receipt date checks before committing a result to the db. This was necessary as I'm not checking if the sample number exists (assumed intensive process), and some results would go in without valid sample/patient.
Menus changed a bit to try and resolve the horrid looping behavior.