Bug fixing
If patientSelectExistingIDsFromList returned False/None, pt_query and sql_result were not scriptable.
Added if pt_query != False to map results to sql_result, or to set both as empty lists (no further tweaks needed for this)

Found that there's no database check on startup. CSV parser failed when mapping analytes to empty returned sql params in getAnalyteIDs. This needs to be fixed, or added clearly to setup.

Commented out the prints in linear_regression's regressSingleForTesting and correlateX