2020 Prediction Category Comparrison Table
New table category_changes to log the last result of a patient against their linear regression predictive value, and the categories they fall into.
db_strings adds CREATE TABLE and INSERT strings, also a query of linear_regression to get that predictive value based on pid
db_methods adds the insert to insertMany and resgressionGetPrediction (which is effectively regressionSelectByPatientAndType but leaves out the join to patient table, so faster?)

Added predictedCategoryChange2020 to linear_regression:
Takes a list of patient IDs to loop, pulls their results of a given analyte and picks the LAST of all results as the most recent. In theory this is just for MDRD / CKD-EPI results so, pass that value to categorise, then get the linear regression calculated prediction value, and categorise that. Save all the data to a table.
This will make pulling SQL summaries easier I think. Quick little exercise before meeting my project supervisor anyway.